Make rrdp and rsync private modules.
