+++ src/metrics.rs
@@ -1,10 +1,12 @@
 //! Monitoring metrics.
 
+use std::{io, process};
 use std::sync::Arc;
+use std::time::{Duration, SystemTimeError};
 use chrono::{DateTime, Utc};
 use log::info;
+use rpki::uri;
 use rpki::tal::TalInfo;
-use crate::{rrdp, rsync};
 
 
 //------------ Metrics -------------------------------------------------------
@@ -18,10 +20,10 @@ pub struct Metrics {
     tals: Vec<TalMetrics>,
 
     /// Rsync metrics.
-    rsync: Vec<rsync::ModuleMetrics>,
+    rsync: Vec<RsyncModuleMetrics>,
 
     /// RRDP metrics.
-    rrdp: Vec<rrdp::ServerMetrics>,
+    rrdp: Vec<RrdpServerMetrics>,
 }
 
 impl Metrics {
@@ -40,14 +42,14 @@ impl Metrics {
 
     pub fn set_rsync(
         &mut self,
-        rsync: Vec<rsync::ModuleMetrics>
+        rsync: Vec<RsyncModuleMetrics>
     ) {
         self.rsync = rsync
     }
 
     pub fn set_rrdp(
         &mut self,
-        rrdp: Vec<rrdp::ServerMetrics>
+        rrdp: Vec<RrdpServerMetrics>
     ) {
         self.rrdp = rrdp
     }
@@ -64,11 +66,11 @@ impl Metrics {
         &self.tals
     }
 
-    pub fn rsync(&self) -> &[rsync::ModuleMetrics] {
+    pub fn rsync(&self) -> &[RsyncModuleMetrics] {
         &self.rsync
     }
 
-    pub fn rrdp(&self) -> &[rrdp::ServerMetrics] {
+    pub fn rrdp(&self) -> &[RrdpServerMetrics] {
         &self.rrdp
     }
 
@@ -129,5 +131,37 @@ impl TalMetrics {
             vrps: 0
         }
     }
+}
+
+
+//------------ RrdpServerMetrics ---------------------------------------------
+
+#[derive(Clone, Debug)]
+pub struct RrdpServerMetrics {
+    pub notify_uri: uri::Https,
+    pub notify_status: Option<reqwest::StatusCode>,
+    pub serial: Option<usize>,
+    pub duration: Result<Duration, SystemTimeError>,
+}
+
+impl RrdpServerMetrics {
+    pub fn new(notify_uri: uri::Https) -> Self {
+        RrdpServerMetrics {
+            notify_uri,
+            notify_status: None,
+            serial: None,
+            duration: Ok(Duration::from_secs(0))
+        }
+    }
+}
+
+
+//------------ RsyncModuleMetrics --------------------------------------------
+
+#[derive(Debug)]
+pub struct RsyncModuleMetrics {
+    pub module: uri::RsyncModule,
+    pub status: Result<process::ExitStatus, io::Error>,
+    pub duration: Result<Duration, SystemTimeError>,
 }
