+++ src/repository/rrdp.rs
@@ -1,7 +1,6 @@
 /// Local repository copy synchronized with RRDP.
 
 pub use self::cache::{Cache, Run, ServerId};
-pub use self::server::ServerMetrics;
 
 mod cache;
 pub mod http;
